Add TLS support
